% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twoCoprimary-package.R
\docType{package}
\name{twoCoprimary-package}
\alias{twoCoprimary}
\alias{twoCoprimary-package}
\title{twoCoprimary: Sample Size and Power Calculation for Two Co-Primary Endpoints}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Comprehensive functions to calculate sample size and power for clinical trials with two co-primary endpoints. The package supports five endpoint combinations: two continuous endpoints (Sozu et al. 2011 \doi{10.1080/10543406.2011.551329}), two binary endpoints using asymptotic methods (Sozu et al. 2010 \doi{10.1002/sim.3972}) and exact methods (Homma and Yoshida 2025 \doi{10.1177/09622802251368697}), mixed continuous and binary endpoints (Sozu et al. 2012 \doi{10.1002/bimj.201100221}), and mixed count and continuous endpoints (Homma and Yoshida 2024 \doi{10.1002/pst.2337}). All methods appropriately account for correlation between endpoints and provide both sample size and power calculation capabilities.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gosukehommaex.github.io/twoCoprimary/}
  \item \url{https://github.com/gosukehommaEX/twoCoprimary}
  \item Report bugs at \url{https://github.com/gosukehommaEX/twoCoprimary/issues}
}

}
\author{
\strong{Maintainer}: Gosuke Homma \email{my.name.is.gosuke@gmail.com} (\href{https://orcid.org/0000-0002-6854-5627}{ORCID})

Authors:
\itemize{
  \item Gosuke Homma \email{my.name.is.gosuke@gmail.com} (\href{https://orcid.org/0000-0002-6854-5627}{ORCID})
}

}
\keyword{internal}
